+++ src/compiler/tree_translator/tree_translator.h
@@ -2,8 +2,8 @@
 #define __TREE_VALIDATOR_H__
 
 #include <stddef.h>
-#include "../config.h"
-#include "../compiler/ast.h"
+#include "../../hardware/config.h"
+#include "../ast.h"
 
 int translate_token_tree(Node* head, byte* result_arr, size_t* result_index);
