+++ src/hardware/config.h
