+++ src/hardware/memory.h
