--- src/instructions.h
@@ -1,12 +1,0 @@
-#ifndef __INSTRUCTIONS_H__
-#define __INSTRUCTIONS_H__
-
-enum INSTRUCTIONS {
-    MEMORY_SECTION = 0x10,
-    LOGIC_SECTION = 0x20,
-    STUFF_SECTION = 0x30,
-    MATH_SECTION = 0xe0,
-    BITWISE_SECTION = 0xf0,
-};
-
-#endif
