+++ src/instructions_executors.h
@@ -2,9 +2,9 @@
 #define __INSTRUCTIONS_EXECUTORS_H__
 
 #include "pc.h"
-#include "config.h"
+#include "hardware/config.h"
 
-/**
+/*
 * @brief Each *_instructions function RUN instruction in PC object
 */
 void memory_instructions(PC* vpc, word code, word op1, word op2);
