+++ src/pc.c
@@ -9,7 +9,7 @@
 
 #include "pc.h"
 #include "instructions_executors.h"
-#include "error_codes.h"
+#include "common/error_codes.h"
 
 void usage(const char* prog_name) {
     fprintf(stderr, "Usage: %s <bin file>\n", prog_name);
